collect_ai_response/gpt53/requirements.txt
openai>=1.40.0
pillow>=10.0.0
pandas
pyarrow
tqdm
ipykernel
